Moving culture specific logic from StringPrototype to StringInlHelper created class.
